% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\docType{data}
\name{dataResults}
\alias{dataResults}
\title{The data.frame containing the results of the simulation study. It is automatically called by the plotting function.}
\format{
A data frame with 53940 rows and 10 variables
}
\usage{
data(dataResults)
}
\description{
The columns of the data.frame are
}
\details{
\itemize{
\item tag. Character vector describing the condition.
\item npcs. Number of principal components used.
\item method. Imputation method used.
\item nla. Number of latent variables used to generate the data.
\item auxcor. Correlation between the main latent variables and the auxiliary ones.
\item pm. Proportion of missing cases on each variable with missing values
\item mech. Missing data mechanism.
\item loc. Location of the missing data in the variable distribution.
\item p. Total number of items.
\item stat. Statistics computed.
\item vars. Variables involved in the statistic.
\item est_avg. Average estimate of the statistic over the Monte Carlo repetitions.
\item mcsd. Standard deviation of the statistic estimate over the Monte Carlo repetitions.
\item ref. True value of the statistic.
\item RB. Raw bias.
\item PRB. Percent relative bias.
\item CIC. Confidence interval coverage.
\item CIW. Average confidence interval width.
\item CIW_sd. Standard deviation of the confidence interval width.
\item CIW_lwr_avg. Average value of the confidence interval lower bound.
\item CIW_upr_avg. Average value of the confidence interval upper bound.
}
}
\keyword{datasets}
