% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mi.spcr-package.R
\docType{package}
\name{plot.mi.spcr-package}
\alias{plot.mi.spcr}
\alias{plot.mi.spcr-package}
\title{plot.mi.spcr: Plot mi-spcr results}
\description{
A shiny app to interact with the results of the mi-spcr study.
}
\author{
\strong{Maintainer}: Edoardo Costantini \email{e.costantini@tilburguniversity.edu} (\href{https://orcid.org/0000-0001-9581-9913}{ORCID})

}
\keyword{internal}
