% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_simulation.R
\name{plot_simulation}
\alias{plot_simulation}
\title{Plot simulation results}
\usage{
plot_simulation(
  results,
  outcome,
  y_axis_range,
  parameter,
  variables,
  n_latent,
  na_mechanism,
  prop_na,
  method_vector,
  npc_range
)
}
\arguments{
\item{results}{object containing results produced by the simulation study}

\item{outcome}{performance measure to plot}

\item{parameter}{estimated parameter of interest}

\item{method_vector}{experimental factor value: imputation methods considered}

\item{npc_range}{experimental factor value: number of components considered}

\item{n_items}{experimental factor value: number of items}

\item{latent_structure}{experimental factor value: whether the latent structure is imposed or not}

\item{categories}{experimental factor value: number of categories of discretized variables}
}
\value{
Returns the ggplot
}
\description{
Generate the main plot for the simulation study.
}
\examples{
# Define example inputs
results <- dataResults
outcome <- "PRB"
n_latent <- dataResults$nla[1]
na_mechanism <- levels(dataResults$mech)
prop_na <- sort(unique(dataResults$pm))
variables <- unique(dataResults$vars)[1]
parameter <- unique(dataResults$stat)[1]
method_vector <- levels(dataResults$method)[1:4]
npc_range <- c(0, 12)
y_axis_range <- c(0, 10)

# Use the function
plot_simulation(
    results = dataResults,
    outcome = "PRB",
    n_latent = dataResults$nla[2],
    na_mechanism = levels(dataResults$mech),
    prop_na = sort(unique(dataResults$pm)),
    variables = unique(dataResults$vars)[1],
    parameter = unique(dataResults$stat)[1],
    method_vector = levels(dataResults$method)[1:4],
    npc_range = c(0, 12),
    y_axis_range = c(0, 40)
)

}
\author{
Edoardo Costantini, 2023
}
