% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{start_app}
\alias{start_app}
\title{start_app}
\usage{
start_app()
}
\value{
Shiny app UI.
}
\description{
Starts a Shiny app to interact with the results of the \href{https://github.com/EdoardoCostantini/mi-spcr}{mi-spcr} project.
}
\details{
The interface of the Shiny app allows you to change the values of the following simulation study experimental factors:
\itemize{
\item Number of latent variables used to generate the data (2, 10, 50)
\item Proportion of missing values imposed on every variable (0.1, 0.25, 0.5)
\item Missing data mechanism imposed (MCAR, MAR)
\item Imputation methods (see names in the interface)
\itemize{
\item pcr: mice with principal component regression as univariate imputation method
\item spcr: mice with supervised principal component regression as univariate imputation method
\item plsr: mice with partial least squares regression as univariate imputation method
\item pcovr: mice with principal covariates regression as univariate imputation method
\item qp: mice with normal linear model with bootstrap as univariate imputation method and quickpred() used to select the predictors
\item am: mice with normal linear model with bootstrap as univariate imputation method and the analysis model variables used as predictors
\item all: mice with normal linear model with bootstrap as univariate imputation method and all available items used as predictors
\item cc: complete case analysis
\item fo: fully observed data (results if there had been no missing values)
}
\item Number of principal components (depending on the number of latent variables used)
\item Outcome measure:
\itemize{
\item RB: raw bias
\item PRB: percent relative bias
\item CIC: confidence interval coverage
\item CIW: average confidence interval
\item mcsd: Standard deviation of the estimate across the monte carlo simulations
}
\item Statistic:
\itemize{
\item cor: correlation between two items with missing values
\item cov: covariance between two items with missing values
\item cor: mean of an item with missing values
\item cor: variance of an item with missing values
}
\item Variables considered
\item Zoom on the y-axis
}
}
